requirements.txt
setuptools
plyfile
pytorch-lightning==2.4.0
opencv-python
imageio
ftfy
regex
tqdm
#./third_party/CLIP/
git+https://github.com/openai/CLIP.git
altair
streamlit
protobuf
timm
tensorboard
tensorboardX
matplotlib
test-tube
wandb
torchmetrics
scikit-image
scikit-learn
#git+https://github.com/zhanghang1989/PyTorch-Encoding/
pycocotools
onnxruntime 
onnx